Fix build on Mac OS X 10.5
Original commit message from CVS:
2007-11-20  Julien MOUTTE  <julien@moutte.net>

* ext/taglib/gsttaglibmux.c: (gst_tag_lib_mux_render_tag),
(gst_tag_lib_mux_adjust_event_offsets):
* gst/qtdemux/qtdemux.c: (qtdemux_parse_theora_extension):
* sys/osxaudio/Makefile.am:
* sys/osxvideo/cocoawindow.h:
* sys/osxvideo/cocoawindow.m: Fix build on Mac OS X 10.5

+++ sys/osxvideo/cocoawindow.h
@@ -35,7 +35,7 @@ struct _GstOSXImage;
 @interface GstGLView : NSOpenGLView
 {
     int i_effect;
-    unsigned long pi_texture;
+    unsigned int pi_texture;
     float f_x;
     float f_y;
     int initDone;
